refactor(app): bundle accept-invitation multi-step in factory + hook
- accept-invitation factory chains accept -> setActive in its mutationFn
  (CLAUDE.md storage cookbook \xc2\xa75 "multi-step workflows go in the factory");
  returns void since organizationId is an internal workflow detail.
- new useAcceptInvitation() feature hook bundles refetch \xc3\x973 (session +
  active-org + orgs-list) + broadcastAuthChange + toast + navigate, mirroring
  the use-set-active-org pattern.
- accept.page drops useState, useNavigate, useQueryClient, two useMutation
  calls, manual try/catch, broadcastAuthChange, toast, and the async onAccept
  wrapper. Page is now ~50 lines, single-purpose.
