+++ apps/app/src/adapters/mutations/accept-invitation.ts
@@ -1,17 +1,9 @@
 import { mutationOptions } from "@tanstack/react-query";
 import { authClient } from "../auth-client";
 
-export interface AcceptInvitationResult {
-  organizationId: string | null;
-}
-
 export const acceptInvitationMutationOptions = mutationOptions({
   mutationKey: ["org", "accept-invitation"] as const,
-  mutationFn: async ({
-    invitationId,
-  }: {
-    invitationId: string;
-  }): Promise<AcceptInvitationResult> => {
+  mutationFn: async ({ invitationId }: { invitationId: string }): Promise<void> => {
     const { data, error } = await authClient.organization.acceptInvitation({ invitationId });
     if (error) throw error;
 
@@ -19,8 +11,13 @@ export const acceptInvitationMutationOptions = mutationOptions({
       invitation?: { organizationId?: string };
       organizationId?: string;
     } | null;
-
     const organizationId = payload?.invitation?.organizationId ?? payload?.organizationId ?? null;
-    return { organizationId };
+
+    if (organizationId) {
+      const { error: setActiveError } = await authClient.organization.setActive({
+        organizationId,
+      });
+      if (setActiveError) throw setActiveError;
+    }
   },
 });
